+++ src/foundation/src/PDFsharp/src/PdfSharp.Charting/Charting.Renderers/AreaChartRenderer.cs
@@ -144,7 +144,7 @@ namespace PdfSharp.Charting.Renderers
                 var elements = sri.Series._seriesElements;
                 if (elements != null)
                 {
-                    sri.PointRendererInfos = new PointRendererInfo[elements.Count ];
+                    sri.PointRendererInfos = new PointRendererInfo[elements.Count];
                     for (int pointIdx = 0; pointIdx < sri.PointRendererInfos.Length; ++pointIdx)
                     {
                         var pri = new PointRendererInfo();
@@ -155,7 +155,7 @@ namespace PdfSharp.Charting.Renderers
                             pri.LineFormat = sri.LineFormat;
                             pri.FillFormat = sri.FillFormat;
                             if (point._lineFormat != null && !point._lineFormat.Color.IsEmpty)
-                                pri.LineFormat = new XPen(point._lineFormat.Color, point._lineFormat.Width.Point);
+                                pri.LineFormat = new XPen(point._lineFormat.Color, (float_)point._lineFormat.Width.Point);
                             if (point._fillFormat != null && !point._fillFormat.Color.IsEmpty)
                                 pri.FillFormat = new XSolidBrush(point._fillFormat.Color);
                         }
